Clean up logging in file writer engine
